存檔點傳送功能

+++ Assets/_Script/Player/PlayerStates/SuperStates/PlayerAbilityState.cs
@@ -94,7 +94,7 @@ public class PlayerAbilityState : PlayerState
         {
             foreach (IMapDamageableItem mapDamageableItem in Combat.DetectedMapDamageableItems.ToList())
             {
-                mapDamageableItem.TakeDamage((int)damageAmount);
+                mapDamageableItem.TakeDamage(damageAmount);
             }
         }
     }
